Fix triggering rules for the "tarball-create" job
Commit fd8ce681890d9db2a35cf3d05d067527e0b8e3d1 (a backport of commit
4d5d3b75da12a6e4c1180bc003d88badf52cb514) did not account for the fact
that the "tarball-create" GitLab CI job is not created for manually
triggered pipelines.  This prevents manual pipeline creation from
succeeding as it causes the "gcc:tarball" job to have unsatisfied
dependencies.  Make sure the "tarball-create" job is created for
manually triggered pipelines to allow such pipelines to be started
again.