Clean up the "installdirs" target in Makefile.in
The isc-config.sh script was removed in 2019 [1], so the top-level
Makefile does not need to create ${bindir} and ${mandir}/man1 any more.

All named options creating files in the ${localstatedir}/run/named
directory are able to automatically create that directory (as well as
its containing directories), so there is no need for the top-level
Makefile to create ${localstatedir}/run.

Clean up the "installdirs" target in the top-level Makefile accordingly.

[1] see commit 4b44351e6514cc4fd46bd0beb9adc5f3f31c481d